+++ lending/src/sdk/models/shared/accountinginvoice.ts
@@ -8,9 +8,9 @@ import { AccountingPaymentAllocation } from "./accountingpaymentallocation";
 import { DataType } from "./datatype";
 import { InvoiceLineItem } from "./invoicelineitem";
 import { InvoiceStatus } from "./invoicestatus";
-import { Items } from "./items";
 import { Metadata } from "./metadata";
 import { SupplementalData } from "./supplementaldata";
+import { WithholdingTaxItems } from "./withholdingtaxitems";
 import { Expose, Type } from "class-transformer";
 
 export class SalesOrderReference extends SpeakeasyBase {
@@ -345,8 +345,8 @@ export class AccountingInvoice extends SpeakeasyBase {
     @Expose({ name: "totalTaxAmount" })
     totalTaxAmount: number;
 
-    @SpeakeasyMetadata({ elemType: Items })
+    @SpeakeasyMetadata({ elemType: WithholdingTaxItems })
     @Expose({ name: "withholdingTax" })
-    @Type(() => Items)
-    withholdingTax?: Items[];
+    @Type(() => WithholdingTaxItems)
+    withholdingTax?: WithholdingTaxItems[];
 }