+++ lending/src/sdk/models/shared/index.ts
@@ -272,3 +272,4 @@ export * from "./transferaccount";
 export * from "./validation";
 export * from "./validationitem";
 export * from "./weblink";
+export * from "./withholdingtaxitems";