+++ lending/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v0.4.0] lending
 ### Releases
 - [NPM v0.4.0] https://www.npmjs.com/package/lending/v/0.4.0 - lending
+
+## 2023-12-25 01:29:55
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.129.1 (2.223.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.1] lending
+### Releases
+- [NPM v0.4.1] https://www.npmjs.com/package/lending/v/0.4.1 - lending